+++ core/src/v2/schema.rs
@@ -130,6 +130,9 @@ pub trait TypedData {
 }
 
 macro_rules! impl_type_simple {
+    ($ty:ty) => {
+        impl TypedData for $ty {}
+    };
     ($ty:ty, $dt:expr) => {
         impl TypedData for $ty {
             fn data_type() -> DataType {
@@ -189,6 +192,8 @@ impl_type_simple!(
     DataType::File,
     DataTypeFormat::Binary
 );
+#[cfg(feature = "actix-session")]
+impl_type_simple!(actix_session::Session);
 #[cfg(feature = "chrono")]
 impl_type_simple!(
     chrono::NaiveDateTime,